Split up IEncounterable generators
Can break the filtering out into smaller methods such that the generator
can return results with less strict filtering
